requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.1.0
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
httpx==0.28.1
boto3==1.38.39
pydantic==2.10.6
aiohttp==3.11.16
